+++ src/WordPress/Shortcodes/ShowPlayersShortcode.php
@@ -1,0 +1,110 @@
+<?php
+/**
+ * OpenTT - Table Tennis Management Plugin
+ * Copyright (C) 2026 Aleksa Dimitrijević
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License,
+ * or (at your option) any later version.
+ */
+
+namespace OpenTT\Unified\WordPress\Shortcodes;
+
+final class ShowPlayersShortcode
+{
+    public static function render($atts, array $deps)
+    {
+        $call = static function ($name, ...$args) use ($deps) {
+            $name = (string) $name;
+            if (!isset($deps[$name]) || !is_callable($deps[$name])) {
+                return null;
+            }
+            return $deps[$name](...$args);
+        };
+
+        $atts = shortcode_atts([
+            'klub' => '',
+        ], $atts);
+
+        $klub_id = 0;
+        if (!empty($atts['klub'])) {
+            $klub_post = get_page_by_path(sanitize_title((string) $atts['klub']), OBJECT, 'klub');
+            if (!$klub_post) {
+                $klub_post = get_page_by_title((string) $atts['klub'], OBJECT, 'klub');
+            }
+            if ($klub_post && !is_wp_error($klub_post)) {
+                $klub_id = intval($klub_post->ID);
+            }
+        } elseif (is_singular('klub')) {
+            $klub_id = intval(get_the_ID());
+        }
+
+        if ($klub_id <= 0) {
+            return '<p>Nije pronađen klub.</p>';
+        }
+
+        $q = new \WP_Query([
+            'post_type' => 'igrac',
+            'posts_per_page' => -1,
+            'meta_query' => [
+                'relation' => 'OR',
+                [
+                    'key' => 'povezani_klub',
+                    'value' => $klub_id,
+                    'compare' => '=',
+                ],
+                [
+                    'key' => 'klub_igraca',
+                    'value' => $klub_id,
+                    'compare' => '=',
+                ],
+            ],
+            'orderby' => 'title',
+            'order' => 'ASC',
+        ]);
+
+        if (!$q->have_posts()) {
+            return (string) $call('shortcode_title_html', 'Igrači') . '<p>Nema registrovanih igrača za ovaj klub.</p>';
+        }
+
+        ob_start();
+        echo (string) $call('shortcode_title_html', 'Igrači');
+        echo '<div class="stoni-igraci-list">';
+        while ($q->have_posts()) {
+            $q->the_post();
+            $id = intval(get_the_ID());
+            $slika = get_the_post_thumbnail($id, 'medium', ['class' => 'stoni-igrac-slika']);
+            if (!$slika) {
+                $slika = '<img src="' . esc_url((string) $call('player_fallback_image_url')) . '" alt="Igrač" class="stoni-igrac-slika" />';
+            }
+            $ime = (string) get_the_title($id);
+            $link = get_permalink($id);
+
+            $ime_ime = $ime;
+            $ime_prezime = '';
+            if (strpos($ime, ' ') !== false) {
+                $parts = explode(' ', $ime, 2);
+                $ime_ime = (string) ($parts[0] ?? '');
+                $ime_prezime = (string) ($parts[1] ?? '');
+            }
+
+            echo '<div class="stoni-igrac-card">';
+            echo '<a href="' . esc_url($link) . '" class="stoni-igrac-row">';
+            echo '<div class="stoni-igrac-left">';
+            echo $slika; // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped
+            echo '<div class="stoni-igrac-ime">';
+            echo '<span class="stoni-igrac-ime-ime">' . esc_html($ime_ime) . '</span>';
+            echo '<span class="stoni-igrac-ime-prezime">' . esc_html($ime_prezime) . '</span>';
+            echo '</div>';
+            echo '</div>';
+            echo '<div class="stoni-igrac-right"><span class="stoni-igrac-detalji">Detalji igrača</span></div>';
+            echo '</a>';
+            echo '</div>';
+        }
+        echo '</div>';
+        wp_reset_postdata();
+
+        return ob_get_clean();
+    }
+}